+++ apps/game/tests/functional/test_results_page.py
@@ -1,12 +1,10 @@
-# coding=UTF-8
-
 from django.contrib.auth import get_user_model
 from django.core.urlresolvers import reverse
 
 from rankme.tests import RankMeTestCase
 
-from ..models import Game
-from .factories import UserFactory, CompetitionFactory
+from ...models import Game
+from ..factories import UserFactory, CompetitionFactory
 
 User = get_user_model()
